# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Halaqaty API
  description: |
    REST API for the Halaqaty Quran memorization circle management platform.

    **Authentication:** Flutter uses Firebase Auth for sign-up, sign-in, password
    management, and ID-token refresh. API requests require a Firebase ID token in
    `Authorization: Bearer <token>` and, after backend session creation, the opaque
    current-device session identifier in `X-Halaqaty-Session-ID`. The backend never
    accepts passwords or issues Firebase ID or refresh tokens.

    **Base URL:** `https://api.halaqaty.app/api/v1`

    **WebSocket:** Real-time events (queue, session, chat) are delivered over WebSocket. See [ws_events.md](ws_events.md) for the full event catalogue.
  version: 1.0.0
  contact:
    name: Halaqaty
    url: https://halaqaty.app
servers:
  - url: https://api.halaqaty.app/api/v1
    description: Production
  - url: https://staging.api.halaqaty.app/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
    SessionId: []
...
paths:
...
  /circles:
    get:
      operationId: listCircles
      description: List circles the authenticated user is a member of
      tags: [Circles]
      summary: List circles the authenticated user is a member of
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [teacher, student, supervisor]
          description: Filter by the user's role in the circle
      responses:
        '200':
          description: List of circles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CircleSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createCircle
      description: |
        Create a new circle. The creator may assign existing registered users as one
        or more teachers and one optional backup supervisor. If no teacher is
        selected, the creator becomes teacher; otherwise the creator becomes
        supervisor.
      tags: [Circles]
      summary: Create a new circle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCircleRequest'
      responses:
        '201':
          description: Circle created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/search:
    get:
      operationId: searchUsers
      description: Search registered users by display name for circle-role selection.
      tags: [Users]
      summary: Search registered users
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
      responses:
        '200':
          description: Matching registered users, limited to 20 results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /circles/{circleId}:
    parameters:
      - $ref: '#/components/parameters/CircleId'
    get:
      operationId: getCircle
      description: Get circle details
      tags: [Circles]
      summary: Get circle details
      responses:
        '200':
          description: Circle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateCircle
      description: Update circle settings (teacher only)
      tags: [Circles]
      summary: Update circle settings (teacher only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCircleRequest'
      responses:
        '200':
          description: Updated circle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: archiveCircle
      description: Archive a circle (teacher only). This is soft-state retirement; hard deletion is prohibited.
      tags: [Circles]
      summary: Archive a circle (teacher only)
      responses:
        '204':
          description: Circle archived
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /circles/{circleId}/join:
    parameters:
      - $ref: '#/components/parameters/CircleId'
    post:
      operationId: joinPublicCircle
      description: Join an active public circle as a student.
      tags: [Circles]
      summary: Join a public circle
      responses:
        '201':
          description: Student membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Circle is private, full, archived, or already joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /circles/discover:
    get:
      operationId: discoverPublicCircles
      description: Discover active public circles using a redacted summary; private and member data are omitted.
      tags: [Circles]
      summary: Discover public circles
      parameters:
        - name: query
          in: query
          schema:
            type: string
            maxLength: 100
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Public circle summaries
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CircleSummary'
                  next_cursor:
                    type: string
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /circles/join:
    post:
      operationId: joinCircle
      description: Join a circle using an invite code
      tags: [Circles]
      summary: Join a circle using an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code:
                  type: string
                  pattern: '^HLQ-[A-HJ-NP-Z2-9]{4}$'
                  example: HLQ-7X2K
      responses:
        '200':
          description: Joined the circle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Circle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Invite code not found or expired
        '409':
          description: Circle is full or archived, the user is already a member, or the five-circle limit is reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /circles/{circleId}/messages:
    parameters:
      - $ref: '#/components/parameters/CircleId'
    get:
      operationId: listCircleMessages
      description: List non-deleted messages visible in the caller's current retained membership period.
      tags: [Chat]
      summary: List recent messages in a circle chat
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: before
          in: query
          schema:
            type: string
            format: uuid
          description: Return messages before this message ID (for pagination)
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMessages'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: sendCircleMessage
      description: Send a text, image, or file message to circle chat
      tags: [Chat]
      summary: Send a text, image, or file message to circle chat
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /circles/{circleId}/messages/search:
    parameters:
      - $ref: '#/components/parameters/CircleId'
    get:
      operationId: searchCircleMessages
      tags: [Chat]
      summary: Search visible circle-message history
      description: Full-text search over non-deleted messages visible in the caller's current retained membership period.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 200
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 100 }
        - name: before
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Matching visible messages
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedMessages' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /circles/{circleId}/messages/pinned:
    parameters:
      - $ref: '#/components/parameters/CircleId'
    get:
      operationId: listPinnedCircleMessages
      tags: [Chat]
      summary: List active pinned circle messages
      description: List at most five non-deleted pinned messages visible in the caller's retained membership period, ordered by pinned_at DESC then id DESC.
      responses:
        '200':
          description: Visible pinned messages
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    maxItems: 5
                    items: { $ref: '#/components/schemas/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase JWT
    SessionId:
      type: apiKey
      in: header
      name: X-Halaqaty-Session-ID
      description: Opaque backend-issued identifier for the current device session.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or revoked Firebase ID token or backend session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Request rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CircleSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        max_capacity:
          type: integer
        gender_restriction:
          type: string
          enum: [male, female, mixed, unspecified]
        language:
          type: string
        created_at:
          type: string
          format: date-time
...
    UserSearchResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/UserSearchResult'
...